FIX README, CHANGE how default redis works
The README was incorrect from the merge with the external project so
that was fixed and it was also updated a lot.

The setupRedis() function was updated to set the option.redis.type
setting to 'default' if the user passed no option.redis at all.
But the default handling of this case remains the same.