Add .flow() to FlowClass
This allows FlowClass to start a flow using .flow() and have it work
just like a job using .job().

Added the ability to mark a flow as a helper by using
.startFlow('type', data, helper).  If helper is true it will be removed
from the Kue queue on restarts, and it is assumed the flow job will be
restarted in some other way, either manually by the user or by the
parent FlowClass that started the flow with a .flow() call.

FlowAPI now passes in its startFlow() to FlowClass because requiring
flowAPI as a whole will cause circular references since FlowAPI needs to
require FlowClass itself.