Make FlowClass prune uncompleted relatedJobs
On startup FlowClass wasn't resetting it's relatedJobs field properly in
Mongo or in its JS instance.  It was continually adding and adding which
increased the number of relatedJobs on every server restart, but never
removed jobs that were added to relatedJobs but were never removed.

So now if the FlowClass is restarting a flow, it will remove all related
jobs that were on steps further than what stepsTaken's value is, and
also will remove any substeps that were added at the current step that
were not set as complete.

What this means is that the results and data of a substep's sibling
jobs will be available on each restart but no bad information (jobs that
were related but never completed) will stick around.

Also fixed was active helper flow jobs were not being removed from the
Kue queue on startup which was also playing a role in why the relatedJobs
field was growing and growing.