Fix FlowClass.relatedJobs to correctly update
The relatedJobs were not being correctly updated with the .flow()
substeps.  Here were the problems:

1. Old relatedJobs were not being correctly removed from Mongo on
startup.  The cleanupRelatedJobs() takes care of that now after .start()
has finished running and before the jobs start to be handled.

2. The .job() and .flow() functions were relying on the 'this' value
being passed to them and having the correct this.relatedJobs.  But this
would not happen because jobs were only attached to this.relatedJobs
after they were enqueued into the Kue queue.  This would happen async
while the .job()/.flow() was called synchronously so their 'this' would
never actually have any relatedJobs.  This was working partially because
they would check Mongo for some of the cases which would hold some
relatedJob info but this was unreliable.

3. Steps and substeps were not being passed to flow substeps so they
had no idea how many steps/substeps they had taken on restart and would
run themselves all over again each time.  Now these are passed correctly
in .flow() so they restart at the correct position.