Change flowAPI, jobAPI to add to FloughInstance
Before this commit the flowAPIs and jobAPIs were attached to the
FloughAPI prototype, which was fine until I needed jobAPIs to share
state.  This was because at job registration time I wanted a function to
be passed that would be tied to the job type that was run at job start
time.  For these functions to be persisted in memory and accessible by
.startJob() in the future these dynamic property functions needed to be
stored on a FloughAPI instance rather than the FloughAPI prototype.