Implement working version of FlowClass.execF()
The first skeleton version of .execF() was not in a working state, this
commit brings it to a working state.  What it now does is add this job
to the .promised step it specifies without putting it through
.handleJob() since it is not technically a job but rather a connecting
function.

This commit also implements the incrementing of stepsTaken from -1 to 0
when the initialization steps are completed.