Make startFlow/startJob clone the data it was handed
Before the data that was passed to these functions was manipulated
directly which would cause the original data object to change if Flough
temporarily messed with the data's properties.  Instead we clone the
data so that the data that was passed in will always stay the same on
the user's end.